Remove some casting warnings on 64bit windows builds.

+++ base/gxicolor.c
@@ -698,11 +698,11 @@ image_render_color_thresh(gx_image_enum *penum_orig, const byte *buffer, int dat
                 xrun += penum->x_extent.x;
             vdi = penum->hci;
             contone_stride = penum->line_size;
-            offset_threshold = (- (((long)(penum->thresh_buffer)) +
-                                      penum->ht_offset_bits)) & 15;
+            offset_threshold = (- (((int)(intptr_t)(penum->thresh_buffer)) +
+                                   penum->ht_offset_bits)) & 15;
             for (k = 0; k < spp_out; k ++) {
-                offset_contone[k]   = (- (((long)(penum->line)) +
-                                          (long)contone_stride * k +
+                offset_contone[k]   = (- (((int)(intptr_t)(penum->line)) +
+                                          contone_stride * k +
                                           penum->ht_offset_bits)) & 15;
             }
             data_length = dest_width;
@@ -725,10 +725,10 @@ image_render_color_thresh(gx_image_enum *penum_orig, const byte *buffer, int dat
             xrun = dda_current(dda_ht);            /* really yrun, but just used here for landscape */
             dest_height = gxht_dda_length(&dda_ht, src_size);
             data_length = dest_height;
-            offset_threshold = (-(long)(penum->thresh_buffer)) & 15;
+            offset_threshold = (-(int)(intptr_t)(penum->thresh_buffer)) & 15;
             for (k = 0; k < spp_out; k ++) {
-                offset_contone[k]   = (- ((long)(penum->line) +
-                                          (long)contone_stride * k)) & 15;
+                offset_contone[k]   = (- ((int)(intptr_t)(penum->line) +
+                                          contone_stride * k)) & 15;
             }
             /* In the landscaped case, we want to accumulate multiple columns
                of data before sending to the device.  We want to have a full
